1000fps snake

+++ src/main/kotlin/org/ksnake/Vec2.kt
@@ -1,6 +1,6 @@
 package org.ksnake
 
 data class Vec2 (
-    var x: Byte,
-    var y: Byte,
+    var x: Short,
+    var y: Short,
 )
